split out NoExposureHistoryScreen
